Move GEOSgcm_GridComp under Components in GEOSldas
Currently, GEOSgcm_GridComp is cloned via mepo under GEOSldas_GridComp (which echoes how it
was in in GEOSldas v17 and older). This PR moves GEOSgcm_GridComp under Components in GEOSldas. This
makes its location consistent with GEOSgcm and GEOSldas.

Note that this PR requires a corresponding change in the XXX repo: